% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_theta_types}
\alias{coerce_theta_types}
\title{Enforce integer parameters if requested}
\usage{
coerce_theta_types(theta, integer_params = NULL)
}
\description{
Enforce integer parameters if requested
}
\keyword{internal}
