% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_stopping.R
\name{generate_test_grid}
\alias{generate_test_grid}
\title{Generate Test Grid for Variance Computation}
\usage{
generate_test_grid(X, n = 200)
}
\arguments{
\item{X}{Design matrix to extract bounds from}

\item{n}{Number of test points}
}
\value{
Matrix of test points (n x d)
}
\description{
Creates a space-filling test grid using Sobol sequences.
}
\keyword{internal}
