% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bo_warmstart.R
\name{load_bo_state}
\alias{load_bo_state}
\title{Load BO State}
\usage{
load_bo_state(path)
}
\arguments{
\item{path}{File path to load}
}
\value{
BO fit object
}
\description{
Loads a previously saved BO fit object.
}
