% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{parse_constraints}
\alias{parse_constraints}
\title{Normalise user-supplied constraint specification}
\usage{
parse_constraints(constraints)
}
\description{
Normalise user-supplied constraint specification
}
\keyword{internal}
