% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{plot_benchmark_efficiency}
\alias{plot_benchmark_efficiency}
\title{Plot simulation efficiency comparison}
\usage{
plot_benchmark_efficiency(benchmark)
}
\arguments{
\item{benchmark}{object returned by [benchmark_methods()].}
}
\value{
ggplot object.
}
\description{
Plot simulation efficiency comparison
}
