% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{plot_benchmark_trajectory}
\alias{plot_benchmark_trajectory}
\title{Plot benchmark trajectories (objective vs evaluation)}
\usage{
plot_benchmark_trajectory(benchmark)
}
\arguments{
\item{benchmark}{object returned by [benchmark_methods()].}
}
\value{
ggplot object.
}
\description{
Plot benchmark trajectories (objective vs evaluation)
}
