% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{plot_case_covariance}
\alias{plot_case_covariance}
\title{Plot covariance matrix of gradient effects}
\usage{
plot_case_covariance(covariance)
}
\arguments{
\item{covariance}{covariance matrix from [case_study_diagnostics()].}
}
\value{
ggplot object.
}
\description{
Plot covariance matrix of gradient effects
}
