% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{plot_case_gradient}
\alias{plot_case_gradient}
\title{Plot case-study gradient heatmap}
\usage{
plot_case_gradient(gradients_tbl)
}
\arguments{
\item{gradients_tbl}{tibble from [case_study_diagnostics()].}
}
\value{
ggplot.
}
\description{
Plot case-study gradient heatmap
}
