% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{plot_case_sobol}
\alias{plot_case_sobol}
\title{Plot Sobol indices for the case study}
\usage{
plot_case_sobol(sobol_tbl)
}
\arguments{
\item{sobol_tbl}{tibble from [case_study_diagnostics()].}
}
\value{
ggplot object.
}
\description{
Plot Sobol indices for the case study
}
