% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{plot_feasible_frontier}
\alias{plot_feasible_frontier}
\title{Feasible frontier plot highlighting regulatory constraints}
\usage{
plot_feasible_frontier(fit)
}
\arguments{
\item{fit}{object returned by [bo_calibrate()].}
}
\value{
ggplot object.
}
\description{
Feasible frontier plot highlighting regulatory constraints
}
