% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_gradient_heatmap}
\alias{plot_gradient_heatmap}
\title{Plot gradient heatmap}
\usage{
plot_gradient_heatmap(gradient_tbl)
}
\arguments{
\item{gradient_tbl}{tibble from [sensitivity_diagnostics()] (`$gradients`).}
}
\value{
ggplot object.
}
\description{
Plot gradient heatmap
}
