% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation.R
\name{plot_multifidelity_tradeoff}
\alias{plot_multifidelity_tradeoff}
\title{Visualise multi-fidelity budget vs accuracy trade-offs}
\usage{
plot_multifidelity_tradeoff(ablation)
}
\arguments{
\item{ablation}{object returned by [ablation_multifidelity()].}
}
\value{
ggplot object.
}
\description{
Visualise multi-fidelity budget vs accuracy trade-offs
}
