% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_sobol_indices}
\alias{plot_sobol_indices}
\title{Plot Sobol indices}
\usage{
plot_sobol_indices(sobol_tbl)
}
\arguments{
\item{sobol_tbl}{tibble returned by [sa_sobol()].}
}
\value{
ggplot object.
}
\description{
Plot Sobol indices
}
