% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{plot_tradeoff_surfaces}
\alias{plot_tradeoff_surfaces}
\title{Trade-off surface plot across key operating characteristics}
\usage{
plot_tradeoff_surfaces(fit)
}
\arguments{
\item{fit}{object returned by [bo_calibrate()].}
}
\value{
ggplot object.
}
\description{
Trade-off surface plot across key operating characteristics
}
