% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bo_parameter_importance.R
\name{print.param_recommendations}
\alias{print.param_recommendations}
\title{Print method for parameter recommendations}
\usage{
\method{print}{param_recommendations}(x, ...)
}
\arguments{
\item{x}{Object of class "param_recommendations"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for parameter recommendations
}
