% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{prob_feasibility}
\alias{prob_feasibility}
\title{Probability of feasibility under Gaussian predictive distribution}
\usage{
prob_feasibility(mean, sd, constraint_tbl)
}
\description{
Probability of feasibility under Gaussian predictive distribution
}
\keyword{internal}
