% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bo_warmstart.R
\name{remove_fixed_from_bounds}
\alias{remove_fixed_from_bounds}
\title{Remove Fixed Parameters from Bounds}
\usage{
remove_fixed_from_bounds(bounds, fixed_params)
}
\arguments{
\item{bounds}{Original bounds}

\item{fixed_params}{Named vector of fixed parameters}
}
\value{
Reduced bounds list
}
\description{
Helper to remove fixed parameters from bounds list.
}
