% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sa_sobol}
\alias{sa_sobol}
\title{Global sensitivity via Saltelli Sobol indices}
\usage{
sa_sobol(surrogates, bounds, outcome = "EN", n_mc = 1000)
}
\arguments{
\item{surrogates}{list of surrogate models (from [fit_surrogates()]).}

\item{bounds}{parameter bounds used to sample the design space.}

\item{outcome}{operating characteristic of interest.}

\item{n_mc}{number of Monte Carlo samples.}
}
\value{
tibble with first-order Sobol indices per parameter.
}
\description{
Global sensitivity via Saltelli Sobol indices
}
