% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_unit}
\alias{scale_to_unit}
\title{Internal helpers for parameter transformations}
\usage{
scale_to_unit(theta, bounds)
}
\description{
Internal helpers for parameter transformations
}
\keyword{internal}
