% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{summarise_benchmark}
\alias{summarise_benchmark}
\title{Summarise benchmark performance across strategies}
\usage{
summarise_benchmark(benchmark)
}
\arguments{
\item{benchmark}{object returned by [benchmark_methods()].}
}
\value{
tibble with strategy-level summaries.
}
\description{
Summarise benchmark performance across strategies
}
