% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_list_to_vec}
\alias{theta_list_to_vec}
\title{Convert theta list to named numeric vector}
\usage{
theta_list_to_vec(theta)
}
\description{
Convert theta list to named numeric vector
}
\keyword{internal}
