% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_to_id}
\alias{theta_to_id}
\title{Create deterministic identifier for theta in unit space}
\usage{
theta_to_id(unit_theta)
}
\description{
Create deterministic identifier for theta in unit space
}
\keyword{internal}
