% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogates.R
\name{predict_surrogates}
\alias{predict_surrogates}
\title{Predict metrics from fitted surrogates}
\usage{
predict_surrogates(surrogates, unit_x)
}
\description{
Predict metrics from fitted surrogates
}
\keyword{internal}
